UsbDk: Avoid multiple redirections of the same device
For some USB devices system calls our AddDevice callback for FDO.
In addition we call the same callback for PDO of the devices we
want to redirect. In this case we will try to install our filter
two times into the same stack which is problematic.

This patch eliminates such a possibility.

Signed-off-by: Dmitry Fleytman <dfleytma@redhat.com>