InstallHelper: optionally suppress message boxes
Add optional command-line parameter to suppress message
box in case the installation requires reboot or aborted.
This supports use case when the UsbDk is installed along
with other product and restart is required regardless
installation status of UsbDk. When installation helper
runs with -in or -iN, the message boxes are suppressed.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>